Revert "Refactor JSON-LD parsers > moved duplicate code to parent class"
This reverts commit 6343cfd587f8cbd82896469e84d8d11c0e1f8ee2.
